refactor: rename package components for consistency and clarity
