+++ apps/web/src/services/ops-flow/components/PackagePanel.vue
